更新

+++ goods/src/main/java/cn/edu/xmu/goods/model/vo/SkuChangeVo.java
@@ -10,6 +10,6 @@ public class SkuChangeVo {
     Long originalPrice;
     String configuration;
     Long weight;
-    Long inventory;
+    Integer inventory;
     String detail;
 }
